Make replacement of fields in bibliography output more flexible
Previously, replacement of placeholders in the formatting string used
the fields of the actual bib entry. This fails/leaves placeholders in
the output whenever a field is in the format string, but not in the
entry. Thus, this commit turns the logic around: The format string is
parsed for placeholders and these are either replaced by the
corresponding value from the entry or removed if the entry does not
contain a corresponding value.

Furthermore, the format string often contains punctuation, e.g.,
commas, between placeholders. These should of course also be removed
when a placeholder cannot be "filled" with data. Hence, this commit
also introduces a slightly more complex syntax for format strings:

- Literals (i.e., stuff that will always be printed) are given as
  usual
- Placeholders are written in all UPPERCASE and enclosed in @, e.g.,
  "@TITLE@".
- Each placeholder must be given inside a group. Groups are delimited
  by braces, and each group contains exactly one placeholder. Groups
  cannot be nested. A group only produces output if its placeholder
  has corresponding data in the currently processed entry.

Example for @inproceedings:
{@AUTHOR@}{ (@YEAR@)}{: "@TITLE@"}{, in: <em>@BOOKTITLE@</em>}{, p. @PAGES@}{, @PUBLISHER@}.{ @NOTE@}{ <a href="https://doi.org/@DOI@">(DOI)</a>}{ <a href="@URL@">(Link)</a>}