Fix adding entries to sqlite DB
With the previous code, a '?' in the inserted values led to sqlite
expecting more parameters. Since the addition was done entry by entry
(instead of one atomic transaction) anyway, this does not impact
performance.

Regarding performance, a test with >2.000 entries worked just fine
(when updating the "database" page, as well as on a page which cited
all these entries).